Mv QuoteInputData to entropy-shared
