+++ crates/shared/Cargo.toml
@@ -15,6 +15,7 @@ serde       ={ version="1.0", default-features=false, features=["derive"] }
 serde_derive="1.0.147"
 strum       ="0.26.3"
 strum_macros="0.26.4"
+blake2      ={ version="0.10.4", default-features=false }
 
 sp-runtime ={ version="32.0.0", default-features=false, optional=true, features=["serde"] }
 sp-std     ={ version="12.0.0", default-features=false }
