+++ crates/shared/src/types.rs
@@ -14,6 +14,7 @@
 
 #![allow(dead_code)]
 use super::constants::VERIFICATION_KEY_LENGTH;
+use blake2::{Blake2b512, Digest};
 #[cfg(not(feature = "wasm"))]
 use codec::alloc::vec::Vec;
 use codec::{Decode, Encode};
@@ -99,3 +100,22 @@ pub enum HashingAlgorithm {
 
 /// A compressed, serialized [synedrion::ecdsa::VerifyingKey<k256::Secp256k1>]
 pub type EncodedVerifyingKey = [u8; VERIFICATION_KEY_LENGTH as usize];
+
+/// Input data to be included in a TDX attestation
+pub struct QuoteInputData(pub [u8; 64]);
+
+impl QuoteInputData {
+    pub fn new(
+        account_id: [u8; 32],
+        x25519_public_key: X25519PublicKey,
+        nonce: [u8; 32],
+        block_number: u32,
+    ) -> Self {
+        let mut hasher = Blake2b512::new();
+        hasher.update(account_id);
+        hasher.update(x25519_public_key);
+        hasher.update(nonce);
+        hasher.update(block_number.to_be_bytes());
+        Self(hasher.finalize().into())
+    }
+}
